--- test/Adafruit_GPS.h
@@ -1,260 +1,0 @@
-/***********************************
-This is the Adafruit GPS library - the ultimate GPS library
-for the ultimate GPS module!
-
-Tested and works great with the Adafruit Ultimate GPS module
-using MTK33x9 chipset
-    ------> http://www.adafruit.com/products/746
-Pick one up today at the Adafruit electronics shop
-and help support open source hardware & software! -ada
-
-Adafruit invests time and resources providing this open source code,
-please support Adafruit and open-source hardware by purchasing
-products from Adafruit!
-
-Written by Limor Fried/Ladyada  for Adafruit Industries.
-BSD license, check license.txt for more information
-All text above must be included in any redistribution
-****************************************/
-// Fllybob added lines 34,35 and 40,41 to add 100mHz logging capability
-
-#ifndef _ADAFRUIT_GPS_H
-#define _ADAFRUIT_GPS_H
-
-#ifdef __AVR__
-  #if ARDUINO >= 100
-    #include <SoftwareSerial.h>
-  #else
-    #include <NewSoftSerial.h>
-  #endif
-#endif
-
-#include <time.h>
-
-// different commands to set the update rate from once a second (1 Hz) to 10 times a second (10Hz)
-// Note that these only control the rate at which the position is echoed, to actually speed up the
-// position fix you must also send one of the position fix rate commands below too.
-#define PMTK_SET_NMEA_UPDATE_100_MILLIHERTZ  "$PMTK220,10000*2F" // Once every 10 seconds, 100 millihertz.
-#define PMTK_SET_NMEA_UPDATE_200_MILLIHERTZ  "$PMTK220,5000*1B"  // Once every 5 seconds, 200 millihertz.
-#define PMTK_SET_NMEA_UPDATE_1HZ  "$PMTK220,1000*1F"
-#define PMTK_SET_NMEA_UPDATE_5HZ  "$PMTK220,200*2C"
-#define PMTK_SET_NMEA_UPDATE_10HZ "$PMTK220,100*2F"
-// Position fix update rate commands.
-#define PMTK_API_SET_FIX_CTL_100_MILLIHERTZ  "$PMTK300,10000,0,0,0,0*2C" // Once every 10 seconds, 100 millihertz.
-#define PMTK_API_SET_FIX_CTL_200_MILLIHERTZ  "$PMTK300,5000,0,0,0,0*18"  // Once every 5 seconds, 200 millihertz.
-#define PMTK_API_SET_FIX_CTL_1HZ  "$PMTK300,1000,0,0,0,0*1C"
-#define PMTK_API_SET_FIX_CTL_5HZ  "$PMTK300,200,0,0,0,0*2F"
-// Can't fix position faster than 5 times a second!
-
-
-#define PMTK_SET_BAUD_57600 "$PMTK251,57600*2C"
-#define PMTK_SET_BAUD_9600 "$PMTK251,9600*17"
-
-// turn on only the second sentence (GPRMC)
-#define PMTK_SET_NMEA_OUTPUT_RMCONLY "$PMTK314,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0*29"
-// turn on GPRMC and GGA
-#define PMTK_SET_NMEA_OUTPUT_RMCGGA "$PMTK314,0,1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0*28"
-// turn on ALL THE DATA
-#define PMTK_SET_NMEA_OUTPUT_ALLDATA "$PMTK314,1,1,1,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0*28"
-// turn off output
-#define PMTK_SET_NMEA_OUTPUT_OFF "$PMTK314,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0*28"
-
-// to generate your own sentences, check out the MTK command datasheet and use a checksum calculator
-// such as the awesome http://www.hhhh.org/wiml/proj/nmeaxor.html
-
-#define PMTK_LOCUS_STARTLOG  "$PMTK185,0*22"
-#define PMTK_LOCUS_STOPLOG "$PMTK185,1*23"
-#define PMTK_LOCUS_STARTSTOPACK "$PMTK001,185,3*3C"
-#define PMTK_LOCUS_QUERY_STATUS "$PMTK183*38"
-#define PMTK_LOCUS_ERASE_FLASH "$PMTK184,1*22"
-#define LOCUS_OVERLAP 0
-#define LOCUS_FULLSTOP 1
-
-#define PMTK_ENABLE_SBAS "$PMTK313,1*2E"
-#define PMTK_ENABLE_WAAS "$PMTK301,2*2E"
-
-// standby command & boot successful message
-#define PMTK_STANDBY "$PMTK161,0*28"
-#define PMTK_STANDBY_SUCCESS "$PMTK001,161,3*36"  // Not needed currently
-#define PMTK_AWAKE "$PMTK010,002*2D"
-
-// ask for the release and version
-#define PMTK_Q_RELEASE "$PMTK605*31"
-
-// request for updates on antenna status
-#define PGCMD_ANTENNA "$PGCMD,33,1*6C"
-#define PGCMD_NOANTENNA "$PGCMD,33,0*6D"
-
-// definitions for EPO Packet field offsets
-#define EPO_PREAMBLE_OFFSET 0
-#define EPO_LENGTH_OFFSET 2
-#define EPO_COMMAND_OFFSET 4
-#define EPO_SEQUENCE_OFFSET 6
-#define EPO_SATELLITE_OFFSET 8
-#define EPO_CHECKSUM_OFFSET 188
-#define EPO_ENDWORD_OFFSET 189
-#define EPO_PACKET_LENGTH 191
-
-// Constants for setting the output mode of the GPS
-#define PMTK_OUTPUT_FORMAT_BINARY 1
-#define PMTK_OUTPUT_FORMAT_NMEA 0
-
-// Constants for PMTK001 Acknowledgement
-#define PMTK_ACK_NO_RESPONSE -1
-#define PMTK_ACK_INVALID 0
-#define PMTK_ACK_UNSUPPORTED 1
-#define PMTK_ACK_FAILED 2
-#define PMTK_ACK_SUCCEEDED 3
-
-
-// how long to wait when we're looking for a response
-#define MAXWAITSENTENCE 5
-
-// Support for Particle.io builds
-#if defined (SPARK)
-  #include "application.h"
-#else
-  #if ARDUINO >= 100
-   #include "Arduino.h"
-  #if defined (__AVR__) && !defined(__AVR_ATmega32U4__)
-   #include "SoftwareSerial.h"
-  #endif
-  #else
-   #include "WProgram.h"
-   #include "NewSoftSerial.h"
-  #endif
-#endif
-
-
-class Adafruit_GPS {
- public:
-  void begin(uint32_t baud);
-
-#if defined(SPARK)
-  Adafruit_GPS(); // Constructor on Particle.io devices
-#else
-  #ifdef __AVR__
-    #if ARDUINO >= 100
-      Adafruit_GPS(SoftwareSerial *ser); // Constructor when using SoftwareSerial
-    #else
-      Adafruit_GPS(NewSoftSerial  *ser); // Constructor when using NewSoftSerial
-    #endif
-  #endif
-    Adafruit_GPS(HardwareSerial *ser); // Constructor when using HardwareSerial
-#endif
-
-  char *lastNMEA(void);
-  boolean newNMEAreceived();
-  void common_init(void);
-
-  void sendCommand(const char *);
-
-  void writePmtkChecksum(char *);
-
-  void pause(boolean b);
-
-  boolean parseNMEA(char *response);
-  uint8_t parseHex(char c);
-
-  void flush(void);
-
-  char read(void);
-  boolean parse(char *);
-  void interruptReads(boolean r);
-
-  boolean wakeup(void);
-  boolean standby(void);
-
-  uint8_t hour, minute, seconds, year, month, day;
-  uint16_t milliseconds;
-  // Floating point latitude and longitude value in degrees.
-  float latitude, longitude;
-  // Fixed point latitude and longitude value with degrees stored in units of 1/100000 degrees,
-  // and minutes stored in units of 1/100000 degrees.  See pull #13 for more details:
-  //   https://github.com/adafruit/Adafruit-GPS-Library/pull/13
-  int32_t latitude_fixed, longitude_fixed;
-  float latitudeDegrees, longitudeDegrees;
-  float geoidheight, altitude;
-  float speed, angle, magvariation, HDOP;
-  char lat, lon, mag;
-  boolean fix;
-  uint8_t fixquality, satellites;
-
-  boolean waitForSentence(const char *wait, uint8_t max = MAXWAITSENTENCE);
-  boolean LOCUS_StartLogger(void);
-  boolean LOCUS_StopLogger(void);
-  boolean LOCUS_ReadStatus(void);
-
-  uint16_t LOCUS_serial, LOCUS_records;
-  uint8_t LOCUS_type, LOCUS_mode, LOCUS_config, LOCUS_interval, LOCUS_distance, LOCUS_speed, LOCUS_status, LOCUS_percent;
-
-  // Functions for sending EPO data
-  bool startEpoUpload();
-  bool sendEpoSatellite(char *);
-  bool endEpoUpload(void);
-
-  // Functions for sending location and time hints to the GPS to speed up the
-  // Time To First Fix (TTFF). Send a time hint before sennding a location hint!
-  long gpsTimeToUTC(long, long);
-  void sendLocationHint(float, float, int, int, int, int, int, int, int);
-  void sendTimeHint(int, int, int, int, int, int);
-  void sendEpoDataRequest();
-
-  // Functions for determining the status of the above commands
-  // (They take a long time to complete execution)
-  int getTimeHintStatus();
-  int getLocationHintStatus();
-  int getEpoDataRequestResponse();
-
-  // Binary packet utility functions, necessary for sending EPO data,
-  // but also useful to drone community. (They seem to like binary
-  // protocols better.)
-  char checksum(char *, int, int);
-  void send_binary_command(uint16_t, char *, int);
-  bool send_buffer(char*, int);
-  bool byte_available(void);
-  char read_byte(void);
-  void format_packet(uint16_t, char *, int, char *);
-  bool waitForPacket(char*, int, long);
-  void format_acknowledge_packet(char*, uint16_t);
-  void set_output_format(int);
-  void dump_binary_packet(void);
-  bool flush_epo_packet(void);
-
-  // During NMEA parsing, if a PMTK001 (Command acknowledgement packet)
-  // is parsed, store the status here
-  int16_t lastMTKAcknowledged;
-  int16_t lastMTKStatus;
-
-  // When sendEpoDataRequest is called, store returned values here
-  long epoStartUTC;
-  long epoEndUTC;
-
-
- private:
-  boolean paused;
-
-  // EPO private variables
-  char packet_buffer[EPO_PACKET_LENGTH];
-  uint16_t epo_sequence_number;
-  int satellite_number;
-
-  uint32_t serial_baud;
-
-
-  uint8_t parseResponse(char *response);
-  #if defined(SPARK)
-    #else
-    #ifdef __AVR__
-      #if ARDUINO >= 100
-        SoftwareSerial *gpsSwSerial;
-      #else
-        NewSoftSerial  *gpsSwSerial;
-      #endif
-    #endif
-    HardwareSerial *gpsHwSerial;
-  #endif
-};
-
-
-#endif
